requirements.txt
#Python==3.13.9
Flask==3.1.2
Werkzeug==3.1.3
mysql-connector-python==9.5.0
python-dotenv==1.1.1